fix: link to mend dashboard from dependency dashboards (#18)
Since this config can not know what repository it is used it, I wrongly
assumed I could drop only the repository part of the path. But this only
works if the user is already logged in to the dashboard. Otherwise it
results in a 500 error page.

By linking to the same host without any path, if you are not logged in
you are asked to login, and have to find the repository logs yourself.
Or use the link from any of the PRs.