chore: bump node & use pnpm (#32)
This PR updates most things to our usual versions and tools for
typescript repos.
With that NodeJS version there is no need for .js files anymore, node
can run .ts files,
which are now type checked as part of PR checks.
It uses pnpm, husky for pre-commit hooks, prettier for formatting, etc.

Several renovate config files needed an automated migration, so I
applied them.
The validator action is no longer using a fixed docker image, but `pnpm
dlx` to run the latest version of renovate for validating configs.

This is a subset of the changes in #31, the "trigger" script and
workflow have been removed.

---------

Co-authored-by: Jannis Morgenstern <jannis.n.morgenstern@gmail.com>