fix: restore `:prImmediate` winning over `security:minimumReleaseAgeNpm`
According to https://github.com/renovatebot/renovate/discussions/39242#discussioncomment-15011518
The order matters since the [v42 of renovate](https://github.com/renovatebot/renovate/releases/tag/42.0.0),
which is stricter about adhering to the minimumReleaseAge setting even for manual triggers.

This is an attempt to restore the previous behavior in a central place,
hopefully restoring the creation of PRs in addition to branches.