fix: restore `:prImmediate` winning over `security:minimumReleaseAgeNpm` (#33)
According to
https://github.com/renovatebot/renovate/discussions/39242#discussioncomment-15011518
The order matters since the [v42 of
renovate](https://github.com/renovatebot/renovate/releases/tag/42.0.0),
which is stricter about adhering to the minimumReleaseAge setting even
for manual triggers.
Which is confirmed when reading [the preset
details](https://docs.renovatebot.com/presets-security/#securityminimumreleaseagenpm),
which changed here:
https://github.com/renovatebot/renovate/commit/03716cd287a3a017580bdb5f248f0ac509b7c3f1

It also explicitly sets the `minimumReleaseAge` for our own packages to
`0 days` instead of `null`, to be more explicit and not rely on any
default value.

This is an attempt to restore the previous behavior in a central place,
hopefully restoring the creation of PRs in addition to branches.