fix: properly configure minimumReleaseAge for npm packages
drops the usage of the preset `security:minimumReleaseAgeNpm`
since it interferes with our way of creating PRs.

It was verified in https://github.com/bettermarks/renovate-config-test-javascript/pull/2
that this properly restores the creation of PRs in our setup.

Updates the readme to reflect the current status of the config.
restored minimumReleaseAge on infra as code package group.

https://github.com/renovatebot/renovate/discussions/39242#discussioncomment-14987608
https://docs.renovatebot.com/key-concepts/minimum-release-age/#minimum-release-age